requirements-dev.txt
bandit==1.7.2
djangorestframework-stubs==3.14.5
django-stubs==4.2.7
drf-yasg-stubs==0.1.4
flake8==7.0.0
flake8-black==0.3.6
flake8-bugbear==24.2.6
flake8-debugger==4.1.2
flake8-import-order==0.18.2
flake8-print==5.0.0
mypy==1.8.0
pep8==1.7.1
pytest-django==4.8.0
pytest-timeout==2.2.0
psycopg2-binary==2.9.9
